separated directories for deploy script
