+++ contracts/deploy/3-aggregator/0-Aggregator.ts
@@ -1,7 +1,7 @@
 import { HardhatRuntimeEnvironment } from 'hardhat/types'
 import { DeployFunction } from 'hardhat-deploy/types'
-import { loadJson } from '../scripts/v0.1/utils'
-import { IAggregatorConfig } from '../scripts/v0.1/types'
+import { loadJson } from '../../scripts/v0.1/utils'
+import { IAggregatorConfig } from '../../scripts/v0.1/types'
 
 const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   const { deployments, getNamedAccounts } = hre
