+++ contracts/package.json
@@ -24,7 +24,10 @@
     "solhint": "./node_modules/.bin/solhint -f table src/**/*.sol",
     "prettier-solidity-write": "./node_modules/.bin/prettier --write src",
     "prettier-solidity-check": "./node_modules/.bin/prettier --check src",
-    "eslint": "DEBUG=eslint:cli-engine npx eslint 'scripts/**'"
+    "eslint": "DEBUG=eslint:cli-engine npx eslint 'scripts/**'",
+    "deploy:rr": "npx hardhat deploy --deploy-scripts deploy/1-requestResponse",
+    "deploy:vrf": "npx hardhat deploy --deploy-scripts deploy/2-vrf",
+    "deploy:aggregator": "npx hardhat deploy --deploy-scripts deploy/3-aggregator"
   },
   "devDependencies": {
     "@ethersproject/abi": "^5.4.7",
